Migrate tests for AppDirsWithDifferentOwner setup check
